+++ src/circuitfp/overlap.py
@@ -2,28 +2,40 @@ import numpy as np
 
 class Overlap:
 
-    def __init__(self, phase1=None, phase2=None, bandwidth=None, gap=None, num_modes=None, basis=None):
-        self.bandwidth = bandwidth
-        self.gap = gap
-        self.num_modes = num_modes
+    def __init__(self, phase1=None, phase2=None, phase_grid=None, bandwidth=None, gap=None, num_modes=None, basis=None):
         self.basis = basis
-        self.phase1 = np.asarray(phase1)
-        self.phase2 = np.asarray(phase2)
-
+        self.phase1 = (
+            np.asarray(phase1)
+            if phase1 is not None
+            else None
+        )
+        self.phase2 = (
+            np.asarray(phase2)
+            if phase2 is not None
+            else None
+        )
+        self.phase_grid = np.asarray(phase_grid) if phase_grid is not None else None
         if basis is not None:
-            self.bandwidth = basis.bandwidth
-            self.gap = basis.gap
-            self.num_modes = basis.num_modes
+            self.bandwidth = float(basis.bandwidth)
+            self.gap = float(basis.gap)
+            self.num_modes = int(basis.num_modes)
+            self.phase_grid = basis.phase_grid
         else:
-            self.bandwidth = bandwidth
-            self.gap = gap
-            self.num_modes = num_modes
+            if bandwidth is None or gap is None or num_modes is None:
+                raise ValueError("If basis is not provided, bandwidth, gap, and num_modes must be specified.")
+            self.bandwidth = float(bandwidth)
+            self.gap = float(gap)
+            self.num_modes = int(num_modes)
 
     
 
     def overlap(self):
         # Calculate the overlap value based on the phase difference
-        phase_difference = np.asarray(self.phase1) - np.asarray(self.phase2)
+        if self.phase_grid is not None:
+            raise ValueError("For overlap calculation, phase1 and phase2 must be provided, not phase_grid.")
+        else:
+            phase_difference = np.asarray(self.phase1) - np.asarray(self.phase2)
+        
         overlap_value = np.exp(self.num_modes * (
             1j * phase_difference / 2
             -
@@ -40,24 +52,13 @@ class Overlap:
     
 
     def overlap_matrix(self):
-        # Check if phase1 and phase2 are 1D arrays
-        if self.phase1.ndim != 1 or self.phase2.ndim != 1:
-            raise ValueError("phase1 and phase2 must be 1D arrays")
+        # Check if phase_grid id 1D arrays
+        if self.phase_grid.ndim != 1:
+            raise ValueError("phase_grid must be 1D arrays")
         
-        # Check if phase1 and phase2 have the same length
-        if self.phase1.shape != self.phase2.shape:
-            raise ValueError(
-                "overlap_matrix() requires phase1 and phase2 to have the same length."
-            )
-        
-        # Check if phase and phase are identical
-        if not np.allclose(self.phase1, self.phase2):
-            raise ValueError(
-                "overlap_matrix() requires phase1 and phase2 to be identical."
-            )
             
         # Create an overlap matrix for the given num_modes in phase space
-        phase_difference = np.asarray(self.phase1)[:, None] - np.asarray(self.phase2)[None, :]
+        phase_difference = np.asarray(self.phase_grid)[:, None] - np.asarray(self.phase_grid)[None, :]
         overlap_matrix = np.exp(self.num_modes * (
             1j * phase_difference / 2
             -